Changed difficulty from int to float
* Now allows non int targets (i.e 31.5)
* Requires a database change
    ALTER TABLE pool_worker MODIFY difficulty float(11);
    ALTER TABLE shares MODIFY difficulty float(11);
* Resolves issue #25
* Updated Schema files

+++ mining/DB_Mysql.py
@@ -309,7 +309,7 @@ class DB_Mysql():
                 "total_rejects": int(data[4]),
                 "total_found": int(data[5]),
                 "alive": True if data[6] is 1 else False,
-                "difficulty": int(data[7])
+                "difficulty": float(data[7])
             }
             
         return ret
